+++ cmd/wardenkms/wardenkms.go
@@ -10,6 +10,8 @@ import (
 	"os"
 	"time"
 
+	"cosmossdk.io/math"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/sethvargo/go-envconfig"
 	"google.golang.org/grpc/connectivity"
 
@@ -28,9 +30,11 @@ type Config struct {
 	KeyringMnemonic string `env:"KEYRING_MNEMONIC, required"`
 	KeyringPassword string `env:"KEYRING_PASSWORD, required"`
 
-	BatchTimeout time.Duration `env:"BATCH_TIMEOUT, default=8s"`
-	BatchSize    int           `env:"BATCH_SIZE, default=20"`
-	GasLimit     uint64        `env:"GAS_LIMIT, default=400000"`
+	BatchInterval time.Duration `env:"BATCH_INTERVAL, default=8s"`
+	BatchSize     int           `env:"BATCH_SIZE, default=7"`
+	GasLimit      uint64        `env:"GAS_LIMIT, default=400000"`
+	TxTimeout     time.Duration `env:"TX_TIMEOUT, default=120s"`
+	TxFee         int64         `env:"TX_FEE, default=400000"`
 
 	HttpAddr string `env:"HTTP_ADDR, default=:8080"`
 
@@ -62,8 +66,10 @@ func main() {
 		Mnemonic:       cfg.Mnemonic,
 		KeychainId:     cfg.KeychainId,
 		GasLimit:       cfg.GasLimit,
-		BatchTimeout:   cfg.BatchTimeout,
+		BatchInterval:  cfg.BatchInterval,
 		BatchSize:      cfg.BatchSize,
+		TxTimeout:      cfg.TxTimeout,
+		TxFees:         sdk.NewCoins(sdk.NewCoin("uward", math.NewInt(cfg.TxFee))),
 	})
 
 	app.SetKeyRequestHandler(func(w keychain.KeyResponseWriter, req *keychain.KeyRequest) {