+++ keychain-sdk/config.go
@@ -3,6 +3,8 @@ package keychain
 import (
 	"log/slog"
 	"time"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
 type Config struct {
@@ -27,9 +29,9 @@ type Config struct {
 	// Mnemonic is the mnemonic to use to derive this Keychain's party private key.
 	Mnemonic string
 
-	// BatchTimeout is the time to wait before sending a batch of requests to the blockchain.
+	// BatchInterval is the time to wait before sending a batch of requests to the blockchain.
 	// Tipically, the batch timeout should be set to the time it takes to produce a block on the blockchain.
-	BatchTimeout time.Duration
+	BatchInterval time.Duration
 
 	// BatchSize is the maximum number of requests to batch together before sending them to the blockchain.
 	BatchSize int
@@ -37,4 +39,13 @@ type Config struct {
 	// GasLimit is the maximum amount of gas to use for each transaction.
 	// The more messages in a batch, the more gas is needed.
 	GasLimit uint64
+
+	// TxFees are the coins used as fees for the transaction sent.
+	TxFees sdk.Coins
+
+	// TxTimeout is the amount of time to wait for a transaction to be included
+	// in a block after having broadcasted it. If the transaction isn't
+	// included in a block, it will be considered as failed (but the blockchain
+	// might still include in a block later).
+	TxTimeout time.Duration
 }