+++ keychain-sdk/keychain.go
@@ -112,10 +112,9 @@ func (a *App) initConnections() error {
 	a.logger().Info("keychain party identity", "address", identity.Address.String())
 
 	txClient := client.NewTxClient(identity, a.config.ChainID, conn, query)
-	a.txWriter = NewTxWriter(txClient, a.config.BatchSize, a.config.BatchTimeout, a.logger())
+	a.txWriter = NewTxWriter(txClient, a.config.BatchSize, a.config.BatchInterval, a.config.TxTimeout, a.logger())
+	a.txWriter.Fees = a.config.TxFees
 	a.txWriter.GasLimit = a.config.GasLimit
 
 	return nil
 }
-
-var defaultPageLimit = uint64(20)